refactor(#2207): move point-read producer routing into producer_point.rs (campsite, #1116)
